+++ CHANGELOG.md
@@ -12,10 +12,10 @@
 
 ### Changed
 * Adjust uniqueness check on ps_component/purl (`OSIDB-4667`)
-
-### Changed
 * Updated aegis feedback method (`AEGIS-203`)
 
+### Added
+* Add KPI metrics data visualization (`AEGIS-196`)
 
 ## [2025.11.4]
 ### Added